+++ lib/comments.php
@@ -2,12 +2,12 @@
 /**
  * Use Bootstrap's media object for listing comments
  *
- * @link http://twitter.github.com/bootstrap/components.html#media
+ * @link http://getbootstrap.com/components/#media
  */
 class Roots_Walker_Comment extends Walker_Comment {
   function start_lvl(&$output, $depth = 0, $args = array()) {
     $GLOBALS['comment_depth'] = $depth + 1; ?>
-    <ul <?php comment_class('media unstyled comment-' . get_comment_ID()); ?>>
+    <ul <?php comment_class('media list-unstyled comment-' . get_comment_ID()); ?>>
     <?php
   }
 
@@ -42,8 +42,10 @@ class Roots_Walker_Comment extends Walker_Comment {
   }
 }
 
-function roots_get_avatar($avatar) {
+function roots_get_avatar($avatar, $type) {
+  if (!is_object($type)) { return $avatar; }
+
   $avatar = str_replace("class='avatar", "class='avatar pull-left media-object", $avatar);
   return $avatar;
 }
-add_filter('get_avatar', 'roots_get_avatar');
+add_filter('get_avatar', 'roots_get_avatar', 10, 2);