+++ lib/config.php
@@ -3,7 +3,6 @@
  * Enable theme features
  */
 add_theme_support('root-relative-urls');    // Enable relative URLs
-add_theme_support('rewrites');              // Enable URL rewrites
 add_theme_support('bootstrap-top-navbar');  // Enable Bootstrap's top navbar
 add_theme_support('bootstrap-gallery');     // Enable Bootstrap's thumbnails component on [gallery]
 add_theme_support('nice-search');           // Enable /?s= to /search/ redirect
@@ -12,8 +11,8 @@ add_theme_support('jquery-cdn');            // Enable to load jQuery from the Go
 /**
  * Configuration values
  */
-define('GOOGLE_ANALYTICS_ID', ''); // UA-XXXXX-Y
-define('POST_EXCERPT_LENGTH', 40); // length in words for excerpt_length filter (ref: http://codex.wordpress.org/Plugin_API/Filter_Reference/excerpt_length)
+define('GOOGLE_ANALYTICS_ID', ''); // UA-XXXXX-Y (Note: Universal Analytics only, not Classic Analytics)
+define('POST_EXCERPT_LENGTH', 40); // Length in words for excerpt_length filter (http://codex.wordpress.org/Plugin_API/Filter_Reference/excerpt_length)
 
 /**
  * .main classes
@@ -21,7 +20,7 @@ define('POST_EXCERPT_LENGTH', 40); // length in words for excerpt_length filter 
 function roots_main_class() {
   if (roots_display_sidebar()) {
     // Classes on pages with the sidebar
-    $class = 'col-sm-8 col-md-9';
+    $class = 'col-sm-8';
   } else {
     // Classes on full width pages
     $class = 'col-sm-12';
@@ -34,7 +33,7 @@ function roots_main_class() {
  * .sidebar classes
  */
 function roots_sidebar_class() {
-  return 'col-sm-4 col-md-3';
+  return 'col-sm-4';
 }
 
 /**
@@ -55,10 +54,8 @@ function roots_display_sidebar() {
      * The second element must be an array even if there's only 1 argument.
      */
     array(
-      'is_404'
-      ,'is_front_page'
-      //array('is_page', array('search'))
-      ,array('is_singular', array('newsletter', 'newsletters'))
+      'is_404',
+      'is_front_page'
     ),
     /**
      * Page template checks (via is_page_template())
@@ -66,11 +63,6 @@ function roots_display_sidebar() {
      */
     array(
       'template-custom.php'
-    , 'template-tutorial.php'
-    , 'template-callout.php'
-    , 'template-application.php'
-    , 'template-one-column.php'
-    , 'template-no-headline.php'
     )
   );
 
@@ -85,13 +77,3 @@ function roots_display_sidebar() {
  * Default: 1140px is the default Bootstrap container width.
  */
 if (!isset($content_width)) { $content_width = 1140; }
-
-/**
- * Define helper constants
- */
-$get_theme_name = explode('/themes/', get_template_directory());
-
-define('RELATIVE_PLUGIN_PATH',  str_replace(home_url() . '/', '', plugins_url()));
-define('RELATIVE_CONTENT_PATH', str_replace(home_url() . '/', '', content_url()));
-define('THEME_NAME',            next($get_theme_name));
-define('THEME_PATH',            RELATIVE_CONTENT_PATH . '/themes/' . THEME_NAME);