+++ lib/custom.php
@@ -2,5 +2,3 @@
 /**
  * Custom functions
  */
-
-include('org/init.php');