+++ templates/footer.php
@@ -1,18 +1,8 @@
-<footer class="content-info container" role="contentinfo">
-  <div class="row">
-    <div class="col-lg-12">
-      <?php dynamic_sidebar('sidebar-footer'); ?>
-      <!--
-      <p>&copy; <?php echo date('Y'); ?> <?php bloginfo('name'); ?></p>
-    	-->
-    </div>
+<footer class="content-info" role="contentinfo">
+  <div class="container">
+    <?php dynamic_sidebar('sidebar-footer'); ?>
+    <p>&copy; <?php echo date('Y'); ?> <?php bloginfo('name'); ?></p>
   </div>
 </footer>
-
-<?php include('org/footer.php'); ?>
-
-<!--[if (gte IE 6)&(lte IE 8)]>
-  <script src="/assets/js/vendor/respond.min.js"></script>
-<![endif]-->
 
 <?php wp_footer(); ?>