+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+### HEAD
+* Update to Bootstrap 3.1.1
+
+### 6.5.2: February 4th, 2014
+* Update to Bootstrap 3.1.0
+* Move DOM routing into an anonymous function to support jQuery noConflict
+* Update to jQuery 1.11.0
+* Add notice to theme activation, tidy activation table markup
+* Remove changing media folder from theme activation (use [Bedrock](https://github.com/roots/bedrock) for clean URLs out of the box)
+* Switch `div.main` to `main` element now that Modernizr uses the latest HTML5 Shiv
+* Update to Modernizr 2.7.0
+* Don't run JSHint on plugins (`assets/js/plugins/`)
+* Disable warnings about undefined variables (JSHint)
+* Merge in updates from HTML5 Boilerplate
+* Add JS source map (disabled by default)
+* Replace `grunt-recess` with `grunt-contrib-less`, add LESS source map support
+
+### 6.5.1: November 5th, 2013
+* Move clean URLs to a [plugin](https://github.com/roots/roots-rewrites)
+* Update to Bootstrap 3.0.1
+
 ### 6.5.0: August 23rd, 2013
 * Reference new site, [http://roots.io/](http://roots.io/)
 * Remove bundled docs, reference [http://roots.io/docs/](http://roots.io/docs/)